Bug 1517398 - Update webrender to commit 075fd68dfcfa2a2fdc66fefe5520caa408adcfcb (WR PR #3438). r=kats
https://github.com/servo/webrender/pull/3438

Differential Revision: https://phabricator.services.mozilla.com/D15648

UltraBlame original commit: 094763412dbf5de18f04c0e37af9a9679b16c73b

+++ gfx/wr/webrender/src/surface.rs
@@ -288,8 +288,8 @@ impl SurfaceDescriptor {
                 
                 
                 
-                let raster_spatial_node = &clip_scroll_tree.spatial_nodes[raster_spatial_node_index.0];
-                let surface_spatial_node = &clip_scroll_tree.spatial_nodes[surface_spatial_node_index.0];
+                let raster_spatial_node = &clip_scroll_tree.spatial_nodes[raster_spatial_node_index.0 as usize];
+                let surface_spatial_node = &clip_scroll_tree.spatial_nodes[surface_spatial_node_index.0 as usize];
 
                 let mut key = CoordinateSpaceMapping::<LayoutPixel, PicturePixel>::new(
                     raster_spatial_node_index,
